memory/tencentdb: consolidate option compatibility tests
